move ZoxModel to resources
